+++ CHANGELOG.md
@@ -32,6 +32,7 @@ TombEngine releases are located in this repository (alongside with Tomb Editor):
 * Fixed motorbike sound going silent while doing jumps.
 * Fixed several original game issues with vehicle collision near crawlspaces and stacked rooms.
 * Fixed ability to draw weapons on a tightrope.
+* Fixed incorrect ladder dismounts onto sloped floors.
 * Fixed dynamic lights disappearing for a single frame after exiting game menus.
 * Fixed controls being stuck when engine window is out of focus.
 * Fixed missing "No" sound when player attempts to interact without particular items in the inventory.